+++ middleware/distributor_skill_test.go
@@ -47,6 +47,8 @@ func newSkillDistributionDB(t *testing.T) *gorm.DB {
 		&skillmodel.Skill{},
 		&skillmodel.SkillVersion{},
 		&skillmodel.UserEnabledSkill{},
+		&skillmodel.SkillPurchaseOrder{},
+		&skillmodel.SkillEntitlement{},
 		&platformmodel.SubscriptionPlan{},
 		&platformmodel.UserSubscription{},
 	); err != nil {